requirements.txt
aiogram>=3.4.1
google-generativeai>=0.4.1
python-dotenv>=1.0.1
Pillow>=10.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
